+++ packages/unified/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @contentmap/unified
 
+## 0.3.0
+
+### Patch Changes
+
+- Updated dependencies
+  - contentmap@0.3.0
+
 ## 0.2.0
 
 ### Patch Changes